requirements.txt
# Clinical Trial Agentic - Dependencies
# Python 3.11+

# Core LLM & Agents
langchain>=0.3.0
langgraph>=0.2.0
langchain-google-genai>=2.0.0
langchain-openai>=0.2.0
google-generativeai>=0.8.0

# Vector Database & Embeddings
chromadb>=0.5.0
sentence-transformers>=3.0.0
# pinecone-client>=3.0.0  # For production

# RAG & Retrieval
rank-bm25>=0.2.2
faiss-cpu>=1.8.0

# API Framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0.0
python-multipart>=0.0.9

# UI
streamlit>=1.40.0
streamlit-extras>=0.4.0
plotly>=5.22.0

# Explainability
shap>=0.45.0
lime>=0.2.0

# Data Processing
pandas>=2.2.0
numpy>=1.26.0

# Medical NLP
spacy>=3.7.0
scispacy>=0.5.4
# en_core_sci_lg  # Install via: pip install https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.4/en_core_sci_lg-0.5.4.tar.gz

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=5.0.0
httpx>=0.27.0

# Documentation
mkdocs>=1.6.0
mkdocs-material>=9.5.0
mkdocstrings[python]>=0.25.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0
tqdm>=4.66.0
loguru>=0.7.0

# Development
black>=24.0.0
ruff>=0.5.0
mypy>=1.10.0
